added exception handling middleware

+++ BusReservations/BusReservations.Core/CommandHandlers/DeleteBusCommandHandler.cs
@@ -1,6 +1,7 @@
 ﻿using BusReservations.Core.Abstract;
 using BusReservations.Core.Commands;
 using BusReservations.Core.Domain;
+using BusReservations.Core.Exceptions;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -22,6 +23,8 @@ namespace BusReservations.Core.CommandHandlers
         public async Task<Bus> Handle(DeleteBusCommand request, CancellationToken cancellationToken)
         {
             var bus = await _unitOfWork.BusRepository.GetBusByID(request.Id);
+            if (bus == null)
+                throw new NotFoundException();
             _unitOfWork.BusRepository.DeleteBus(bus);
             await _unitOfWork.SaveChangesAsync();
             return bus;
